## Summary
* SWIG interfaces for Step in the reference implementation and JIT compiled implementation for the ELL module.
* Python wrappers for Compute around Step when input is synchronous.
* Updated CNTK importer to attach Source, Sink and Clock nodes to imported models.  This is the least intrusive place I can think of.

## Callback plumbing
- Reference implementation: set callbacks by index for a given Source and Sink node. The index is assumed to be fixed by the order of node traversal.  Callbacks can be functions that are set at execution time.
- Compile + JIT implementation is limited to:
  - Only 1 callback per node type is accepted.  However, we can support multiple callbacks per node type by expanding the callback functions with a context parameter (e.g. string).  This work is covered in #1131.
  - Callbacks are well-known named C functions whose names are specified during Compile, or by the map at import time.
  - Future: I investigated setting callbacks by function pointers, and it'll require a bit more plumbing. Specifically, by emitting "Setter" functions that receives a function pointer type for each Source / Sink node.
- For Python, glue code is added to wrap callables (e.g. Python functions) into the respective CallbackBase implementation.  This is unchanged from the V1 implementation of Step.

Related work items: #1042, #1082, #1088, #1106

+++ interfaces/python/package/ell/pretrained_model.py
@@ -89,7 +89,8 @@ class PretrainedModel:
         if not os.path.exists(cmakefile) or _is_file_newer(inpath + '.ell', cmakefile) \
            or not os.path.exists(outpath + '.bc'):
             
-            compiled = ellmap.Compile(target, self.name, 'predict', True)
+            compiled = ellmap.Compile(target, self.name, 'predict', True,
+                dtype=np.float32)
             compiled.WriteBitcode(outpath + '.bc')
             compiled.WriteSwigInterface(outpath + '.i')
