+++ interfaces/python/test/compiled_model_test.py
@@ -2,11 +2,13 @@ from __future__ import print_function
 import functools
 import ell
 import os
+import numpy as np
 
 def test():
     # Load the map created by proton trainer test and compile it
     map = ell.model.Map("protonnTestData.ell")
-    compiledMap = map.Compile("host", "protonn", "predict", False)
+    compiledMap = map.Compile("host", "protonn", "predict", False,
+        dtype=np.float)
     compiledMap.WriteBitcode("protonnTestData.bc");
     
     if os.path.isfile("protonnTestData.bc"):
